+++ src/batch_ops.jl
@@ -58,7 +58,9 @@ Base.@constprop :aggressive function new_entities!(
     components::Tuple;
     relations::Tuple{Vararg{Pair{DataType,Entity}}}=(),
 ) where {F}
-    if n == 0
+    if n < 0
+        throw(ArgumentError("can't add a negative number of entities."))
+    elseif n == 0
         return
     end
     if components isa Tuple{Vararg{DataType}}
@@ -82,7 +84,9 @@ Base.@constprop :aggressive function new_entities!(
     components::Tuple;
     relations::Tuple{Vararg{Pair{DataType,Entity}}}=(),
 )
-    if n == 0
+    if n < 0
+        throw(ArgumentError("can't add a negative number of entities."))
+    elseif n == 0
         return
     end
     rel_types = ntuple(i -> Val(relations[i].first), length(relations))
@@ -559,8 +563,9 @@ end
 
     has_fn = HFN == Val{true}
     return quote
-        _check_locked(world)
+        _check_relation_targets(world, targets)
 
+        _check_locked(world)
         _lock(world._lock)
 
         arches, arches_hot = _get_archetypes(world, filter)
@@ -667,6 +672,7 @@ end
     _check_is_subset(rel_types, add_types)
 
     return quote
+        _check_relation_targets(world, targets)
         _check_locked(world)
         _lock(world._lock)
 
@@ -998,6 +1004,7 @@ end
     world_has_rel = Val{_has_relations(CS)}()
 
     exprs = []
+    push!(exprs, :(_check_relation_targets(world, targets)))
     push!(exprs, :(_check_locked(world)))
     push!(
         exprs,