+++ src/unchecked.jl
@@ -17,7 +17,7 @@ function _add_unchecked!(expr)
     fn_uncheck = (
         :has_components, :get_components, :set_components!, :add_components!,
         :remove_components!, :exchange_components!, :get_relations, :set_relations!,
-        :remove_entity!, :copy_entity!,
+        :remove_entity!, :copy_entity!, :new_entity!,
     )
     fn_uncheck_base = (
         getindex = :(Ark._unchecked_getindex),
@@ -79,13 +79,14 @@ end
 """
     @unchecked block
 
-Removes some checks performed by these functions and their shortcuts:
-`has_components`, `get_components`, `set_components!`, `add_components!`,
-`remove_components!`, `exchange_components!`, `get_relations`, `set_relations!`,
-`remove_entity!`, `copy_entity!`.
+Removes some checks performed by these functions: `has_components`, `get_components`,
+`set_components!`, `add_components!`, `remove_components!`, `exchange_components!`,
+`get_relations`, `set_relations!`, `new_entity`, `remove_entity!`, `copy_entity!`.
 
-In particular, checks about the aliveness of the entity and presence of components
-on which the functions operate are skipped.
+In particular, these checks are skipped:
+
+  - aliveness of the entity and its relation targets;
+  - presence of components on which the functions operate.
 
 !!! warning