+++ CHANGELOG.md
@@ -16,6 +16,11 @@
 - Many batch operations are up to 3x faster thanks to simd (#575)
 - Many single entity operations are now 10-20% faster thanks to table caching in transitions (#555)
 
+### Bugfixes
+
+- Fixes bugs related to relationship being accepted even for stale entities with recycled ids (#605, #604)
+- Fixes a bug where removing multiple relation targets of the same entity caused an error (#611, fixes #606)
+
 ## [[v0.4.0]](https://github.com/ark-ecs/Ark.jl/compare/v0.3.2...v0.4.0)
 
 ### Breaking changes